Kotlinx. `Object` & `Property Descriptor`
